# pkg/server/api/docs/swagger.yaml
components:
  schemas:
...
    bind.ZoneRecord:
      properties:
        name:
          type: string
        priority:
          type: integer
        ttl:
          type: integer
        type:
          type: string
        value:
          type: string
      required:
      - name
      - type
      - value
      type: object
...
    dhcp.leaseView:
      properties:
        duid:
          type: string
        expire:
          type: integer
        hostname:
          type: string
        hw-address:
          type: string
        ip-address:
          type: string
        reserved:
          type: boolean
        service:
          type: string
        state:
          type: integer
        subnet-id:
          type: integer
      required:
      - duid
      - expire
      - hostname
      - hw-address
      - ip-address
      - reserved
      - service
      - state
      - subnet-id
      type: object
...
    dhcp.serviceState:
      properties:
        running:
          type: boolean
        service:
          type: string
      required:
      - running
      - service
      type: object
    dhcp.statsResponse:
      properties:
        services:
          items:
            $ref: '#/components/schemas/dhcp.serviceState'
          type: array
          uniqueItems: false
        stats4:
          items:
            $ref: '#/components/schemas/kea.Stat'
          type: array
          uniqueItems: false
        stats6:
          items:
            $ref: '#/components/schemas/kea.Stat'
          type: array
          uniqueItems: false
      type: object
...
    dns.ddnsDeleteRequest:
      properties:
        name:
          type: string
        type:
          type: string
        value:
          type: string
      required:
      - name
      - type
      type: object
    dns.ddnsZoneView:
      properties:
        error:
          type: string
        name:
          type: string
        records:
          items:
            $ref: '#/components/schemas/bind.ZoneRecord'
          type: array
          uniqueItems: false
      required:
      - name
      type: object
...
    kea.Stat:
      properties:
        name:
          type: string
        value:
          type: number
      required:
      - name
      - value
      type: object
...
    types.ConfigureConntrackRequest:
      properties:
        allow_inbound:
          type: boolean
        friend_address:
          type: string
        friend_interface:
          type: string
        local_address:
          type: string
        local_interface:
          type: string
        port:
          type: integer
      required:
      - friend_address
      - friend_interface
      - local_address
      - local_interface
      type: object
    types.ConfigureConntrackResult:
      properties:
        status:
          type: string
      required:
      - status
      type: object
...
    types.FriendInfo:
      properties:
        enabled:
          type: boolean
        fingerprint:
          type: string
        ha:
          type: boolean
        hostname:
          type: string
        manage:
          type: boolean
        name:
          type: string
        paired:
          type: boolean
        url:
          type: string
      required:
      - enabled
      - ha
      - name
      - url
      type: object
...
    types.FriendStatus:
      properties:
        conntrackd_running:
          type: boolean
        encryption:
          type: boolean
        fingerprint:
          type: string
        hostname:
          type: string
        identity_match:
          type: boolean
        last_error:
          type: string
        last_seen:
          type: string
        name:
          type: string
        os:
          type: string
        reachable:
          type: boolean
        rtt_ms:
          type: integer
        version:
          type: string
        vrrp:
          items:
            $ref: '#/components/schemas/types.FriendVRRPRole'
          type: array
          uniqueItems: false
      required:
      - encryption
      - identity_match
      - name
      - reachable
      type: object
...
    types.FriendVRRPRole:
      properties:
        instance:
          type: string
        state:
          type: string
      required:
      - instance
      - state
      type: object
...
    types.Response-array_dhcp_leaseView:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          items:
            $ref: '#/components/schemas/dhcp.leaseView'
          type: array
          uniqueItems: false
        status:
          type: string
      type: object
    types.Response-array_dns_ddnsZoneView:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          items:
            $ref: '#/components/schemas/dns.ddnsZoneView'
          type: array
          uniqueItems: false
        status:
          type: string
      type: object
...
    types.Response-dhcp_statsResponse:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/dhcp.statsResponse'
        status:
          type: string
      type: object
...
    types.Response-string:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          type: string
        status:
          type: string
      type: object
...
    types.Response-types_ConfigureConntrackResult:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.ConfigureConntrackResult'
        status:
          type: string
      type: object
...
    types.Response-types_FriendInfo:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.FriendInfo'
        status:
          type: string
      type: object
...
    types.Response-types_FriendStatus:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.FriendStatus'
        status:
          type: string
      type: object
...
    types.Response-types_StatusResponse:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/types.StatusResponse'
        status:
          type: string
      type: object
...
    types.StatusResponse:
      properties:
        status:
          type: string
      required:
      - status
      type: object
...
    types.VerifyFriendRequest:
      properties:
        fingerprint:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: '"Bearer <token>"'
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: Configuration and monitoring API for a Routier node.
  title: Routier API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/dhcp/free-ip:
    get:
      parameters:
      - description: subnet cidr or an address inside it
        in: query
        name: target
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_StatusResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: Suggest a free address in a subnet's pools
      tags:
      - dhcp
  /api/dhcp/leases:
    get:
      parameters:
      - description: filter (exact IP is looked up via kea, otherwise substring match
          on ip/mac/duid/hostname)
        in: query
        name: q
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-array_dhcp_leaseView'
          description: OK
      security:
      - BearerAuth: []
      summary: List active DHCP leases, optionally filtered
      tags:
      - dhcp
...
  /api/dhcp/leases/stream:
    get:
      parameters:
      - description: Kea daemon (omitted streams DHCPv4 and DHCPv6)
        in: query
        name: source
        schema:
          enum:
          - kea-dhcp4
          - kea-dhcp6
          - kea-dhcp-ddns
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                type: string
          description: OK
      security:
      - BearerAuth: []
      summary: Tail Kea DHCP daemon logs (SSE)
      tags:
      - dhcp
...
  /api/dhcp/stats:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-dhcp_statsResponse'
          description: OK
      security:
      - BearerAuth: []
      summary: DHCP server state and Kea statistics
      tags:
      - dhcp
...
  /api/dns/ddns:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-array_dns_ddnsZoneView'
          description: OK
      security:
      - BearerAuth: []
      summary: Live DDNS-managed zones and the dynamic records they currently hold
      tags:
      - dns
  /api/dns/ddns/{zone}/records:
    delete:
      parameters:
      - description: zone name
        in: path
        name: zone
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dns.ddnsDeleteRequest'
                description: record to delete
                summary: request
        description: record to delete
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-string'
          description: OK
      security:
      - BearerAuth: []
      summary: Clear a single dynamic record from a DDNS-managed zone
      tags:
      - dns
...
  /api/friends/{name}/check:
    post:
      parameters:
      - description: friend name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_FriendStatus'
          description: OK
      security:
      - BearerAuth: []
      summary: Force a health check + info refresh for one friend
      tags:
      - friends
...
  /api/friends/{name}/conntrack:
    post:
      parameters:
      - description: friend name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.ConfigureConntrackRequest'
              description: conntrack config
              summary: body
        description: conntrack config
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_ConfigureConntrackResult'
          description: OK
      security:
      - BearerAuth: []
      summary: Configure conntrackd HA with a friend (applies on both sides)
      tags:
      - friends
...
  /api/friends/{name}/pair:
    post:
      parameters:
      - description: friend name
        in: path
        name: name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/types.VerifyFriendRequest'
              description: operator-confirmed fingerprint
              summary: body
        description: operator-confirmed fingerprint
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-types_FriendInfo'
          description: OK
      security:
      - BearerAuth: []
      summary: Validate a friend's fingerprint and complete pairing
      tags:
      - friends
...
servers:
- url: /